Fix C++ code practices and reorder includes
It is best practice to only #include the bare minimum
in .cpp files and not .h files. Headers should only include
the bare minimum of other headers. It's also best practice
to forward declare classes and structs instead of including
and entire header file to deal with it. Some resources will
say to use header files in header files, but it increases
compile time objectively.